requirements.txt
selenium
pandas
python-telegram-bot
asyncio
dotenv